requirements.txt
protobuf==3.6.1
tqdm==4.14.0
python-snappy==0.5.3
PyYAML==4.2b1
